Add .local dir to Garmin/ConnectIQ
This is so we can make use of the Gnome Keyring after `docker-compose --rm run`
after a successful login. Users can opt to pick the .local dir in
`$HOME/.Garmin/ConnectIQ` or pick their own homedir. Or pick some other
mechanism. I prefer the former, as it keeps everything in one place.

Signed-off-by: Wesley Schwengle <wesleys@opperschaap.net>